Added code for relay switch.
Added code for relay switch. For now, this just switches the solar panel output on / off into the charge controller. This gets rid of the problem of sometimes getting too much voltage from the output of the solar charge controller (ex: 14 V output is too much for the electronic speed controller, which just beeps without turning anything when the voltage is that high).